fix: move ErrorBoundary component to ErrorBoundary folder
