+++ src/App.tsx
@@ -1,7 +1,7 @@
 import { Component } from 'react';
 import SearchSection from './components/SearchSection';
 import CardSection from './components/CardSection/CardSection';
-import ErrorBoundary from './components/ErrorBoundary';
+import ErrorBoundary from './components/ErrorBoundary/ErrorBoundary';
 import ErrorButton from './components/ErrorButton';
 import Loader from './components/Loader';
 import { fetchCharacters } from './api/api';
